requirements.txt
# requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
google-genai==0.3.0
pydantic==2.6.4
PyYAML==6.0.1
python-dotenv==1.0.1
streamlit==1.32.2
